+++ lib/torch/device.rb
@@ -1,5 +1,11 @@
 module Torch
   class Device
+    def inspect
+      extra = ", index: #{index.inspect}" if index != -1
+      "device(type: #{type.inspect}#{extra})"
+    end
+    alias_method :to_s, :inspect
+
     def ==(other)
       eql?(other)
     end